+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Released
 
 All notable changes to this project will be documented in this file. This project does adhere to [Semantic Versioning](http://semver.org/)
+## 2.3.10 (5/29/17)
+- CORE-7269: Fix Happy Path Military Email IVA test
 
 ## 2.3.9 (5/25/17)
 - CORE-7504: Fix Happy Path Teacher Doc IVA test
@@ -71,6 +73,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ##2.2.92 (4/7/17)
 - CORE-7203: Military Error Codes Tests
+- IDME-3536: Fix HLP broken tests
 
 ##2.2.91 (4/6/17)
 - CORE-7228: Fix logic for clearing out the old test data after dupe tests